fix: resolved issues in treasury contracts
