+++ src/constants/tables/treasuryContractsTableOptions.ts
@@ -1,0 +1,11 @@
+import type { TreasuryContractsColumns } from "@/types/tableTypes";
+
+export const treasuryContractsTableOptions: {
+  key: keyof TreasuryContractsColumns;
+}[] = [
+  { key: "project" },
+  { key: "vendor" },
+  { key: "budget" },
+  { key: "milestones" },
+  { key: "status" },
+];
